HomeScreen

+++ lib/constants.dart
@@ -1,7 +1,7 @@
 class ThemeConstants {
   static final _padding = 16.0;
 
-  static double spacing(int val) {
+  static double spacing(double val) {
     return _padding * val;
   }
 }
